Renamed missing pot evidence rule and test
